requirements.txt
wykop-sdk==0.3.1
requests==2.10.0
SQLObject==2.2.1
